requirements.txt
python-dotenv==1.0.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
redis==5.0.1
gigachat==0.1.43